added reset flow counts
