+++ src/components/floweditor/FlowEditor.tsx
@@ -1,175 +1,31 @@
-import React, { useEffect, useState } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
 import { useMutation, useLazyQuery, useQuery } from '@apollo/client';
 import { Prompt, Redirect, useHistory } from 'react-router-dom';
 import { IconButton } from '@material-ui/core';
-import * as Manifest from '@glific/flow-editor/build/asset-manifest.json';
 
 import { ReactComponent as HelpIcon } from 'assets/images/icons/Help.svg';
 import { ReactComponent as FlowIcon } from 'assets/images/icons/Flow/Dark.svg';
 import { ReactComponent as WarningIcon } from 'assets/images/icons/Warning.svg';
 import { ReactComponent as ExportIcon } from 'assets/images/icons/Flow/Export.svg';
+import { ReactComponent as ResetFlowIcon } from 'assets/images/icons/Flow/ResetFlow.svg';
 import { Button } from 'components/UI/Form/Button/Button';
-import {
-  APP_NAME,
-  FLOW_EDITOR_CONFIGURE_LINK,
-  FLOW_EDITOR_API,
-  FLOWS_HELP_LINK,
-  CONTACT_CHAT_LINK,
-} from 'config/index';
+import { APP_NAME, FLOWS_HELP_LINK } from 'config/index';
 import { Simulator } from 'components/simulator/Simulator';
 import { DialogBox } from 'components/UI/DialogBox/DialogBox';
 import { setNotification } from 'common/notification';
-import { PUBLISH_FLOW } from 'graphql/mutations/Flow';
+import { PUBLISH_FLOW, RESET_FLOW_COUNT } from 'graphql/mutations/Flow';
 import { EXPORT_FLOW, GET_FLOW_DETAILS, GET_FREE_FLOW } from 'graphql/queries/Flow';
 import { setAuthHeaders } from 'services/AuthService';
+import { SideDrawerContext } from 'context/session';
 import { GET_ORGANIZATION_SERVICES } from 'graphql/queries/Organization';
 import { Loading } from 'components/UI/Layout/Loading/Loading';
 import { exportFlowMethod } from 'common/utils';
 import styles from './FlowEditor.module.css';
+import { checkElementInRegistry, loadfiles, setConfig } from './FlowEditor.helper';
 
 declare function showFlowEditor(node: any, config: any): void;
 
-// function to suppress the error for custom registery in floweditor
-const safeDecorator = (fn: any) =>
-  // eslint-disable-next-line
-  function (...args: any) {
-    try {
-      // @ts-ignore
-      return fn.apply(this, args);
-    } catch (error) {
-      if (
-        error instanceof DOMException &&
-        error.message.includes('has already been used with this registry')
-      ) {
-        return false;
-      }
-      throw error;
-    }
-  };
-
-customElements.define = safeDecorator(customElements.define);
-
-const loadfiles = (startFlowEditor: any) => {
-  const files: Array<HTMLScriptElement | HTMLLinkElement> = [];
-  const filesToLoad: any = Manifest.files;
-  let index = 0;
-  Object.keys(filesToLoad).forEach((fileName) => {
-    if (filesToLoad[fileName].startsWith('/static')) {
-      if (filesToLoad[fileName].endsWith('.js')) {
-        index += 1;
-        const script = document.createElement('script');
-        if (index === 2) {
-          script.onload = () => {
-            startFlowEditor();
-          };
-        }
-        script.src = filesToLoad[fileName];
-        script.id = `flowEditorScript${index}`;
-        script.async = false;
-        files.push(script);
-      }
-
-      if (filesToLoad[fileName].endsWith('.css')) {
-        const link = document.createElement('link');
-        link.href = filesToLoad[fileName];
-        link.id = `flowEditorfile${index}`;
-        link.rel = 'stylesheet';
-        document.body.appendChild(link);
-      }
-    }
-  });
-
-  // loading the largest file first
-  document.body.appendChild(files[3]);
-  document.body.appendChild(files[0]);
-  document.body.appendChild(files[2]);
-  document.body.appendChild(files[1]);
-
-  return files;
-};
-
-const glificBase = FLOW_EDITOR_API;
-
-const setConfig = (uuid: any) => ({
-  flow: uuid,
-  flowType: 'messaging',
-  localStorage: true,
-  mutable: true,
-  showNodeLabel: false,
-  attachmentsEnabled: false,
-  filters: ['whatsapp', 'classifier'],
-
-  excludeTypes: [
-    'add_contact_urn',
-    'send_email',
-    'set_run_result',
-    'call_resthook',
-    'start_session',
-    'open_ticket',
-    'transfer_airtime',
-    'split_by_random',
-    'split_by_scheme',
-  ],
-
-  excludeOperators: [
-    'has_text',
-    'has_number_lt',
-    'has_number_lte',
-    'has_number_gte',
-    'has_number_gt',
-    'has_date',
-    'has_date_category',
-    'has_date_lt',
-    'has_number_lte',
-    'has_number_gte',
-    'has_number_gt',
-    'has_date',
-    'has_date_category',
-    'has_date_lt',
-    'has_date_eq',
-    'has_date_gt',
-    'has_time',
-    'has_group',
-    'has_category',
-    'has_state',
-    'has_state_category',
-    'has_district',
-    'has_ward',
-    'has_error',
-    'has_value',
-  ],
-  help: {
-    legacy_extra: 'help.html',
-    missing_dependency: 'help.html',
-    invalid_regex: 'help.html',
-  },
-  endpoints: {
-    simulateStart: false,
-    simulateResume: false,
-    globals: `${glificBase}globals`,
-    groups: `${glificBase}groups`,
-    fields: `${glificBase}fields`,
-    labels: `${glificBase}labels`,
-    channels: `${glificBase}channels`,
-    classifiers: `${glificBase}classifiers`,
-    ticketers: `${glificBase}ticketers`,
-    resthooks: `${glificBase}resthooks`,
-    templates: `${glificBase}templates`,
-    languages: `${glificBase}languages`,
-    attachments: `${glificBase}flow-attachment`,
-    environment: `${glificBase}environment`,
-    recipients: `${glificBase}recipients`,
-    completion: `${glificBase}completion`,
-    activity: `${glificBase}activity`,
-    flows: `${glificBase}flows`,
-    recents: `${glificBase}recents/`,
-    revisions: `${glificBase}revisions/${uuid}`,
-    editor: FLOW_EDITOR_CONFIGURE_LINK,
-    validateMedia: `${glificBase}validate-media`,
-    interactives: `${glificBase}interactive-templates`,
-    contact: CONTACT_CHAT_LINK,
-  },
-});
+customElements.define = checkElementInRegistry(customElements.define);
 
 export interface FlowEditorProps {
   match: any;
@@ -187,6 +43,7 @@ export const FlowEditor = (props: FlowEditorProps) => {
   const [published, setPublished] = useState(false);
   const [stayOnPublish, setStayOnPublish] = useState(false);
   const [modalVisible, setModalVisible] = useState(false);
+  const [showResetFlowModal, setShowResetFlowModal] = useState(false);
   const [lastLocation, setLastLocation] = useState<Location | null>(null);
   const [confirmedNavigation, setConfirmedNavigation] = useState(false);
   const [flowValidation, setFlowValidation] = useState<any>();
@@ -194,6 +51,7 @@ export const FlowEditor = (props: FlowEditorProps) => {
   const [flowKeyword, setFlowKeyword] = useState('');
   const [currentEditDialogBox, setCurrentEditDialogBox] = useState(false);
   const [dialogMessage, setDialogMessage] = useState('');
+  const { drawerOpen } = useContext(SideDrawerContext);
 
   let modal = null;
   let dialog = null;
@@ -237,6 +95,15 @@ export const FlowEditor = (props: FlowEditorProps) => {
       exportFlowMethod(exportData, flowTitle);
     },
   });
+  const [resetFlowCountMethod] = useMutation(RESET_FLOW_COUNT, {
+    onCompleted: ({ resetFlowCount }) => {
+      const { success } = resetFlowCount;
+      if (success) {
+        setNotification('Flow count reset successfully', 'success');
+        setShowResetFlowModal(false);
+      }
+    },
+  });
 
   const [publishFlow] = useMutation(PUBLISH_FLOW, {
     onCompleted: (data) => {
@@ -248,6 +115,24 @@ export const FlowEditor = (props: FlowEditorProps) => {
       }
     },
   });
+
+  const { data: flowName } = useQuery(GET_FLOW_DETAILS, {
+    fetchPolicy: 'network-only',
+    variables: {
+      filter: {
+        uuid,
+      },
+      opts: {},
+    },
+  });
+
+  let flowId: any;
+
+  // flowname can return an empty array if the uuid present is not correct
+  if (flowName && flowName.flows.length > 0) {
+    flowTitle = flowName.flows[0].name;
+    flowId = flowName.flows[0].id;
+  }
 
   const closeModal = () => {
     setModalVisible(false);
@@ -290,22 +175,29 @@ export const FlowEditor = (props: FlowEditorProps) => {
     );
   }
 
-  const { data: flowName } = useQuery(GET_FLOW_DETAILS, {
-    fetchPolicy: 'network-only',
-    variables: {
-      filter: {
-        uuid,
-      },
-      opts: {},
-    },
-  });
+  const handleResetFlowCount = () => {
+    resetFlowCountMethod({ variables: { flowId } });
+  };
 
-  let flowId: any;
-
-  // flowname can return an empty array if the uuid present is not correct
-  if (flowName && flowName.flows.length > 0) {
-    flowTitle = flowName.flows[0].name;
-    flowId = flowName.flows[0].id;
+  if (showResetFlowModal) {
+    modal = (
+      <DialogBox
+        title="Warning!"
+        handleOk={handleResetFlowCount}
+        handleCancel={() => setShowResetFlowModal(false)}
+        colorOk="secondary"
+        buttonOk="Accept & reset"
+        buttonCancel="DON'T RESET YET"
+        alignButtons="center"
+        contentAlign="center"
+        additionalTitleStyles={styles.DialogTitle}
+      >
+        <div className={styles.DialogContent}>
+          Please be careful, this cannot be undone. Once you reset the flow counts you will lose
+          tracking of how many times a node was triggered for users.
+        </div>
+      </DialogBox>
+    );
   }
 
   useEffect(() => {
@@ -459,6 +351,7 @@ export const FlowEditor = (props: FlowEditorProps) => {
       }
     }
   };
+
   return (
     <>
       {dialog}
@@ -529,8 +422,11 @@ export const FlowEditor = (props: FlowEditorProps) => {
       <Prompt when message={handleBlockedNavigation} />
 
       <div className={styles.FlowContainer}>
-        <div className={styles.FlowName} data-testid="flowName">
-          {flowName ? (
+        <div
+          className={drawerOpen ? styles.FlowName : styles.FlowNameClosed}
+          data-testid="flowName"
+        >
+          {flowName && (
             <>
               <IconButton disabled className={styles.Icon}>
                 <FlowIcon />
@@ -538,7 +434,16 @@ export const FlowEditor = (props: FlowEditorProps) => {
 
               {flowTitle}
             </>
-          ) : null}
+          )}
+        </div>
+
+        <div
+          className={drawerOpen ? styles.ResetFlow : styles.ResetClosedDrawer}
+          data-testid="resetFlow"
+          onClick={() => setShowResetFlowModal(true)}
+          aria-hidden="true"
+        >
+          <ResetFlowIcon /> Reset flow counts
         </div>
         <div id="flow" />
         {loading && <Loading />}
